requirements.txt
Django==4.0.5
djangorestframework==3.13.1
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
django-cors-headers==3.7.0
cryptography==38.0.4
aliyun-python-sdk-core-v3==2.13.33
django_oss_storage